requirements.txt

boto3==1.26.160
#ics==0.7.2
icalendar==5.0.7
python-gitlab==3.15.0
pytz==2023.3

aiosmtpd==1.4.4
dkimpy==1.1.5
pyspf==2.0.14
dmarc==1.0.2

aiosqlite==0.19.0
aiohttp==3.8.5